+++ utils/general.py
@@ -1,0 +1,11 @@
+# This module contains general functions
+
+def separate_symbol_and_trade_against(symbol):
+    if symbol.endswith("BTC"):
+        symbol_only = symbol[:-3]
+        symbol_stable = symbol[-3:]
+    elif symbol.endswith(("BUSD","USDT")):    
+        symbol_only = symbol[:-4]
+        symbol_stable = symbol[-4:]
+
+    return symbol_only, symbol_stable